10hz acce, hrm, data prj, with background mode

+++ fetchIMU/inc/bluetooth/gatt/characteristic.h
@@ -3,7 +3,7 @@
 
 bool create_gatt_characteristic();
 bool add_gatt_descriptor_to_gatt_characteristic();
-bool set_gatt_characteristic_value(int, int, int, int);
+bool set_gatt_characteristic_value(int, int16_t, int16_t, int16_t);
 bool notify_gatt_characteristic_value_changed();
 bool get_gatt_characteristic_handle(bt_gatt_h *handle);
